+++ lib/src/country.dart
@@ -3,6 +3,7 @@ import 'dart:math';
 import 'package:flutter/foundation.dart';
 import 'data.dart';
 
+/// Defines length rules for phone numbers for specific country.
 @immutable
 abstract class LengthRule {
   factory LengthRule.range(int min, int max) => _RangeLengthRule(min, max);
@@ -13,33 +14,49 @@ abstract class LengthRule {
   int get maxLength;
 }
 
+/// Holds country related data for phone numbers.
 @immutable
 class Country {
   Country(this.name, this.code, this.prefix, this.length)
       : _prefixStr = prefix.toString(),
         prefixLength = prefix.toString().length;
 
-  final String name;
-  final String code;
-  final int prefix;
-  final int prefixLength;
-  final LengthRule length;
-  final String _prefixStr;
-
-  bool matches(String normalizedNumber) =>
-      normalizedNumber.startsWith(_prefixStr);
-
-  bool isValidNationalNumber(String nationalNumber) =>
-      length.test(nationalNumber.length);
-
-  bool isValidNumber(String normalizedNumber) =>
-      matches(normalizedNumber) &&
-      length.test(normalizedNumber.length - _prefixStr.length);
-
+  /// Returns [Country] instance using [code].
   static Country fromCode(String code) {
     code = code.toUpperCase();
     return countries.firstWhere((c) => c.code == code);
   }
+
+  /// Country name
+  final String name;
+
+  /// Country 2 char alpha code
+  final String code;
+
+  /// Country calling prefix code
+  final int prefix;
+
+  /// Length of the [prefix]
+  final int prefixLength;
+
+  /// Length rule for national number
+  final LengthRule length;
+
+  final String _prefixStr;
+
+  /// Does given [normalizedNumber] matches this country.
+  bool matches(String normalizedNumber) =>
+      normalizedNumber.startsWith(_prefixStr);
+
+  /// Validates given [nationalNumber] using defined [length] rule.
+  bool isValidNationalNumber(String nationalNumber) =>
+      length.test(nationalNumber.length);
+
+  /// Checks whether or not given [normalizedNumber] is valid phone number in
+  /// this country.
+  bool isValidNumber(String normalizedNumber) =>
+      matches(normalizedNumber) &&
+      length.test(normalizedNumber.length - _prefixStr.length);
 }
 
 class _RangeLengthRule implements LengthRule {